image_prediction/src/pipeline.py
import torch
import open_clip
from PIL import Image
import numpy as np
from pathlib import Path
import logging

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class FoodImageAnalyzer:
    def __init__(self, model_name="ViT-B-32", pretrained="laion2b_s34b_b79k"):
        """
        Initialise l'analyseur d'images de nourriture.
        
        Args:
            model_name (str): Nom du modèle OpenCLIP à utiliser
            pretrained (str): Version pré-entraînée du modèle
        """
        self.device = "cuda" if torch.cuda.is_available() else "cpu"
        logger.info(f"Utilisation du device: {self.device}")
        
        # Chargement du modèle OpenCLIP
        self.model, _, self.preprocess = open_clip.create_model_and_transforms(
            model_name, pretrained=pretrained
        )
        self.model = self.model.to(self.device)
        
    def preprocess_image(self, image_path):
        """
        Prétraite l'image pour l'analyse.
        
        Args:
            image_path (str): Chemin vers l'image
            
        Returns:
            torch.Tensor: Image prétraitée
        """
        try:
            image = Image.open(image_path).convert('RGB')
            return self.preprocess(image).unsqueeze(0).to(self.device)
        except Exception as e:
            logger.error(f"Erreur lors du prétraitement de l'image: {e}")
            raise
            
    def extract_features(self, image_tensor):
        """
        Extrait les caractéristiques de l'image avec OpenCLIP.
        
        Args:
            image_tensor (torch.Tensor): Image prétraitée
            
        Returns:
            torch.Tensor: Embeddings de l'image
        """
        with torch.no_grad():
            image_features = self.model.encode_image(image_tensor)
            image_features /= image_features.norm(dim=-1, keepdim=True)
        return image_features
    
    def predict_nutrition(self, image_path):
        """
        Prédit les valeurs nutritionnelles d'une image de nourriture.
        
        Args:
            image_path (str): Chemin vers l'image
            
        Returns:
            dict: Prédictions nutritionnelles
        """
        try:
            # Prétraitement
            image_tensor = self.preprocess_image(image_path)
            
            # Extraction des caractéristiques
            features = self.extract_features(image_tensor)
            
            # TODO: Implémenter la prédiction des valeurs nutritionnelles
            # Pour l'instant, retournons des valeurs fictives
            predictions = {
                "calories": 0.0,
                "proteines": 0.0,
                "lipides": 0.0,
                "glucides": 0.0
            }
            
            return predictions
            
        except Exception as e:
            logger.error(f"Erreur lors de la prédiction: {e}")
            raise

def main():
    # Exemple d'utilisation
    analyzer = FoodImageAnalyzer()
    # TODO: Ajouter un exemple de prédiction avec une image de test
    
if __name__ == "__main__":
    main() 