image_prediction/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
open_clip_torch>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.65.0 